Feature: Finish last Crud endpoints.

+++ FinanzasGrupo2API/DataFrances/Resources/DataFrancesResource.cs
@@ -14,7 +14,7 @@
         public int PlazoGracia { get; set; }
         public float Capital { get; set; }
         public float TeSemestral { get; set; }
-        public int CreditoCapitalizado { get; set; }
+        public float CreditoCapitalizado { get; set; }
         public float NuevaCuota { get; set; }
 
         //Relationships
